Make the generated QR be self-sufficient
Instead of externally hosting the QR on a pomf instance as previous, we upload the QR as an attachment named "qr.png" to the message and embed the image attachment://qr.png. This alllows it all to stay in a tightly knit bundle on Discord itself. This also removes the need for pomf.rb.